override default layout to add js
The theme we're assuming is https://github.com/pages-themes/primer.
Following its instructions, we've copied the master
_layouts/default.html and modified it to add a link to assets/js/main.js
at the bottom of the body (so that the DOM will be ready).

We use the new main.js to find all links to uchicago.box.com and set
target="_blank" so that they'll open in a new window.